code cleanup and organization
