+++ src/_scss/pages/state/spendingovertime/chart/barChart.scss
@@ -4,8 +4,8 @@ svg.bar-graph {
     .chart-legend-item {
         .key-label {
             font-size: rem(12);
-            color: #666;
-            fill: #666;
+            color: $gray-60;
+            fill: $gray-60;
         }
     }
 }
