+++ src/js/components/state/spendingovertime/StateTimeVisualizationChart.jsx
@@ -16,135 +16,13 @@ import {
     Legend
 } from 'recharts';
 import { LoadingMessage, NoResultsMessage } from "data-transparency-ui";
-import { formatMoneyWithUnitsShortLabel } from "../../../helpers/moneyFormatter";
-import BarChartLegend from "../../sharedComponents/timeChart/chart/BarChartLegend";
+import CustomShape from "./chartCustomizations/CustomShape";
+import CustomXTick from "./chartCustomizations/CustomXTick";
+import CustomYTick from "./chartCustomizations/CustomYTick";
+import CustomTooltip from "./chartCustomizations/CustomTooltip";
+import CustomLegend from "./chartCustomizations/CustomLegend";
 
-const CustomShapePropTypes = {
-    x: PropTypes.number,
-    y: PropTypes.number,
-    width: PropTypes.number,
-    height: PropTypes.number,
-    focusBar: PropTypes.bool
-};
-
-const CustomShape = (props) => {
-    const {
-        x, y, width, height, focusBar
-    } = props;
-
-    const fill = "#0081a1"; // $cyan-50v
-    let fillOpacity = "1";
-    if (focusBar && !props?.isActive) {
-        fillOpacity = "0.5";
-    }
-
-    const maxWidth = width > 120 ? 120 : width;
-    const translateX = x + ((width / 2) - (maxWidth / 2));
-
-    return (
-        <rect
-            x={translateX}
-            y={height < 0 ? y - Math.abs(height) : y}
-            width={maxWidth}
-            height={Math.abs(height)}
-            fill={fill}
-            fillOpacity={fillOpacity}
-            className="recharts-bars" />
-    );
-};
-
-CustomShape.propTypes = CustomShapePropTypes;
-
-const CustomXTickPropTypes = {
-    x: PropTypes.number,
-    y: PropTypes.number,
-    payload: PropTypes.object
-};
-
-const CustomXTick = (props) => {
-    const {
-        x, y, payload
-    } = props;
-
-    return (
-        <g transform={`translate(${x},${y})`}>
-            <text
-                x={0}
-                y={0}
-                dx={12}
-                dy={12}
-                textAnchor="end"
-                fill="#5C5C5C"
-                fontSize={12}
-                width="40px">
-                {payload.value}
-            </text>
-        </g>);
-};
-
-CustomXTick.propTypes = CustomXTickPropTypes;
-
-const CustomYTickPropTypes = {
-    x: PropTypes.number,
-    y: PropTypes.number,
-    payload: PropTypes.object
-};
-
-const CustomYTick = (props) => {
-    const {
-        x, y, payload
-    } = props;
-
-    return (
-        <g transform={`translate(${x},${y})`}>
-            <text x={0} y={0} dy={0} textAnchor="end" fill="#5C5C5C" fontSize={12} width="48px">
-                {formatMoneyWithUnitsShortLabel(payload.value)}
-            </text>
-        </g>);
-};
-
-CustomYTick.propTypes = CustomYTickPropTypes;
-
-const CustomTooltipPropTypes = {
-    active: PropTypes.bool,
-    payload: PropTypes.array,
-    label: PropTypes.string,
-    onSetFocusBar: PropTypes.func,
-    onMouseLeave: PropTypes.func
-};
-
-const CustomTooltip = (props) => {
-    const {
-        active,
-        payload,
-        label,
-        onSetFocusBar,
-        onMouseLeave
-    } = props;
-
-    if (active && payload && payload.length) {
-        onSetFocusBar(label);
-        return (
-            <div className="custom-tooltip" role="status" aria-live="assertive">
-                <div className="tooltip__title">
-                    {label}
-                </div>
-                <div className="tooltip__text">
-                    <div className="tooltip__text-label">Obligations</div>
-                    <div className="tooltip__text-amount">
-                        {formatMoneyWithUnitsShortLabel(payload[0].value)}
-                    </div>
-                </div>
-            </div>);
-    }
-
-    onMouseLeave();
-    return null;
-};
-
-CustomTooltip.propTypes = CustomTooltipPropTypes;
-
-const StateTimeVisualizationChartPropTypes = {
+const stateTimeVisualizationChartPropTypes = {
     data: PropTypes.object,
     loading: PropTypes.bool
 };
@@ -178,23 +56,6 @@ const StateTimeVisualizationChart = (props) => {
         setFocusBar(state.label);
     };
 
-    const barChartLegnedConfig = [
-        {
-            color: "#0081a1",
-            label: 'Obligations',
-            offset: 0
-        }
-    ];
-
-    const CustomLegend = () => (
-        <svg className="bar-graph" height={20}>
-            <g className="legend-container">
-                <BarChartLegend legend={barChartLegnedConfig} />
-            </g>
-        </svg>
-
-    );
-
     const renderChart = () => {
         if (loading) {
             return <LoadingMessage />;
@@ -225,17 +86,13 @@ const StateTimeVisualizationChart = (props) => {
                         onMouseLeave={onMouseLeave} />
                     <Legend
                         align="left"
-                        iconType="circle"
-                        content={<CustomLegend />}
-                        wrapperStyle={{ left: 60 }}
-                        margin={{
-                            top: 19, right: 50, bottom: 0, left: 8
-                        }} />
+                        content={<CustomLegend barColor="#0081a1" label="Obligations" />}
+                        wrapperStyle={{ left: 60, bottom: 0 }} />
                     <ReferenceLine y={0} stroke="#dfe1e2" />
                     <Bar
                         dataKey="value"
-                        shape={<CustomShape focusBar={focusBar} />}
-                        activeBar={<CustomShape isActive focusBar={focusBar} />}
+                        shape={<CustomShape focusBar={focusBar} barColor="#0081a1" />}
+                        activeBar={<CustomShape isActive focusBar={focusBar} barColor="#0081a1" />}
                         onMouseEnter={onMouseMove}
                         onMouseOut={onMouseLeave}
                         onMouseLeave={onMouseLeave} />
@@ -251,5 +108,5 @@ const StateTimeVisualizationChart = (props) => {
     );
 };
 
-StateTimeVisualizationChart.propTypes = StateTimeVisualizationChartPropTypes;
+StateTimeVisualizationChart.propTypes = stateTimeVisualizationChartPropTypes;
 export default StateTimeVisualizationChart;
